requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
boto3==1.27.1
botocore==1.30.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.9
dnspython==2.3.0
docx2txt==0.8
Flask==2.3.2
Flask-Cors==4.0.0
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
loguru==0.7.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.0
packaging==23.1
pydantic==1.10.11
pypdf==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.18
tenacity==8.2.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
yarl==1.9.2
